Initial commit
This version compiles, but won't run because of circular dependencies. Also, seveal features are disabled, since they could not be easily ported.
This initial commit is only to start with a TS/JS port that is as close to the original Java, before it is refactored to better fit TS/JS design patterns.